pyproject.toml
[project]
name = "guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "litellm>=1.80.10",
    "python-dotenv>=1.2.1",
]
